+++ src/components/keyboard/Key.js
@@ -7,11 +7,11 @@ export default function Key(props) {
     const gameCtx = useContext(GameContext);
 
     function handleClick() {
-        if (gameCtx.gameover === false && (props.letter !== "EN" && props.letter !== "BS") && gameCtx.letterCount < 5) {
+        if (gameCtx.gameover === false && gameCtx.isPaused === false && (props.letter !== "EN" && props.letter !== "BS") && gameCtx.letterCount < 5) {
             gameCtx.addLetter(props.letter);
-        } else if (gameCtx.gameover === false && props.letter === "BS" && gameCtx.letterCount > 0) {
+        } else if (gameCtx.gameover === false && gameCtx.isPaused === false && props.letter === "BS" && gameCtx.letterCount > 0) {
             gameCtx.deleteLetter();
-        } else if (gameCtx.gameover === false && props.letter === "EN") {
+        } else if (gameCtx.gameover === false && gameCtx.isPaused === false && props.letter === "EN") {
             if (gameCtx.letterCount !== 5) {
                 alert('All letters must be filled');
                 return;
